fix: vnPay max amount
fix vnpay max amount to 200 million

+++ backend/Helpers/Payment_src/core/Payment.Service/Vnpay/Response/VnpayPayResponse.cs
@@ -21,7 +21,7 @@ namespace backend.Payment_src.core.Payment.Service.Vnpay.Response
         public string vnp_TxnRef { get; set; } = string.Empty;
         public string vnp_SecureHashType { get; set; } = string.Empty;
         public string vnp_SecureHash { get; set; } = string.Empty;
-        public int? vnp_Amount { get; set; }
+        public Int64? vnp_Amount { get; set; }
         public string? vnp_ResponseCode { get; set; }
         public string vnp_PayDate { get; set; } = string.Empty;
